+++ release_notes.md
@@ -10,7 +10,9 @@
 
 ### Enhancements:
 
-* Lorem ipsum.
+* Added support for encryption
+* Adding SharedGroup::wait_for_change() which allows a thread to sleep until
+  the database changes.
 
 -----------